Implemented checkbox module
